passim: refine behavior description for edge case
The implemented behavior is to fail when a test fails, and otherwise
pass.  The documented behavior could have been interpreted to specify
failure when no tests were provided, which was not the intent.

Add a section describing changes so people can let go of concerns about
why a trivial package like this keeps getting updated.